refactor: 시설 배치가능 영역 구분
- 공방/사냥터에 배치 가능한 시설을 분리

+++ Assets/01_Scripts/Building/PlacedBuilding.cs
@@ -21,7 +21,7 @@ public class PlacedBuilding : MonoBehaviour, IBuildingUIModel
 
     public BuildingDataSO Data { get; private set; }
     public Vector3Int OriginCell { get; private set; }
-    public int RotationIndex { get; private set; }
+    public short RotationIndex { get; private set; }
 
     // 해당 빌딩이 차지하고 있는 셀
     private readonly List<Vector3Int> occupiedCells = new();
@@ -44,7 +44,7 @@ public class PlacedBuilding : MonoBehaviour, IBuildingUIModel
     /// <param name="originCell">배치 셀 위치</param>
     /// <param name="rotationIndex">회전 인덱스 0 ~ 3</param>
     /// <param name="cells">점유하는 셀</param>
-    public void Initialize(BuildingDataSO data, Vector3Int originCell, int rotationIndex, IEnumerable<Vector3Int> cells) 
+    public void Initialize(BuildingDataSO data, Vector3Int originCell, short rotationIndex, IEnumerable<Vector3Int> cells) 
     {
         Data = data;
         OriginCell = originCell;
